Remove context be more simple stupid
