+++ src/components/common/CoreAppWrapper/CoreAppWrapper.tsx
@@ -4,7 +4,6 @@ import SiteNavigation from 'components/Navbar/SiteNavigation'
 import useMobile from 'hooks/Mobile'
 import { useRouter } from 'next/router'
 import { ArcxProvider } from 'providers/ArcxProvider'
-import { EnsCacheProvider } from 'providers/EnsCacheProvider'
 import { EtherPriceProvider } from 'providers/EtherPriceProvider'
 import LanguageProvider from 'providers/LanguageProvider'
 import ReactQueryProvider from 'providers/ReactQueryProvider'
@@ -33,9 +32,7 @@ export const AppWrapper: React.FC = ({ children }) => {
               <ThemeProvider>
                 <EtherPriceProvider>
                   <ArcxProvider>
-                    <EnsCacheProvider>
-                      <_Wrapper>{children}</_Wrapper>
-                    </EnsCacheProvider>
+                    <_Wrapper>{children}</_Wrapper>
                   </ArcxProvider>
                 </EtherPriceProvider>
               </ThemeProvider>
