+++ src/hooks/ensName.ts
@@ -1,40 +1,79 @@
 import { SECONDS_IN_DAY } from 'constants/numbers'
-import { EnsCacheContext } from 'contexts/ensCacheContext'
+import { readProvider } from 'constants/readProvider'
 import { resolveAddress } from 'lib/ens/resolver'
-import { cacheEnsRecord } from 'providers/EnsCacheProvider'
-import { useContext, useEffect } from 'react'
+import { useEffect, useState } from 'react'
+
+export type EnsCacheRecord = {
+  name: string | null
+  expires: number
+}
+
+const ensLocalstorageKey = () =>
+  readProvider?.network?.chainId
+    ? `jb_ensDict_${readProvider.network.chainId}`
+    : undefined
+
+export const getEnsCache = () => {
+  if (typeof window === 'undefined') return {}
+
+  try {
+    return JSON.parse(
+      window.localStorage.getItem(ensLocalstorageKey() ?? '') ?? '{}',
+    ) as Record<string, EnsCacheRecord>
+  } catch (e) {
+    console.warn('ENS storage not found', e)
+    return {}
+  }
+}
+
+export const cacheEnsRecord = (address: string, record: EnsCacheRecord) => {
+  const key = ensLocalstorageKey()
+  if (!key) return
+
+  window.localStorage?.setItem(
+    key,
+    JSON.stringify({
+      ...getEnsCache(),
+      [address]: record,
+    }),
+  )
+}
+
+const getEnsNameFromCache = (address: string, expiryTime: number) => {
+  const cache = getEnsCache()
+  // Try getting from cache first
+  const record = cache[address]
+  if (record?.expires > expiryTime) {
+    return record.name
+  }
+}
 
 /**
  * Try to resolve an address to an ENS name.
  */
 export function useEnsName(address: string | undefined) {
-  const { cache, pending, setCache, setPending } = useContext(EnsCacheContext)
-
-  const now = new Date().valueOf()
+  const [ensName, setEnsName] = useState<string | null>()
+  const [loading, setLoading] = useState<boolean>(false)
 
   useEffect(() => {
-    const getEnsNameFromCache = (address: string) => {
-      // Try getting from cache first
-      const record = cache[address]
-      if (record?.expires > now) {
-        return record.name
-      }
-
-      // if a fetch is already in progress, bail
-      if (pending[address]) {
-        return
-      }
-    }
+    const now = new Date().valueOf()
 
     async function loadEnsName() {
       if (!address) return
 
-      const ensNameFromCache = getEnsNameFromCache(address)
+      const ensNameFromCache = getEnsNameFromCache(address, now)
       // if cache hit, don't fetch
-      // if a fetch is already in progress, bail
-      if (ensNameFromCache !== undefined || pending[address]) return
+      if (ensNameFromCache !== undefined) {
+        setEnsName(ensNameFromCache)
+        return
+      }
 
-      setPending?.({ ...pending, [address]: true })
+      // if a fetch is already in progress, bail
+      if (loading) {
+        return
+      }
+
+      setLoading(true)
       const name = await resolveAddress(address)
 
       const newRecord = {
@@ -43,16 +82,16 @@ export function useEnsName(address: string | undefined) {
       }
 
       // update state
-      setCache?.({ ...cache, [address]: newRecord })
-      setPending?.({ ...pending, [address]: false })
+      setEnsName(name)
+      setLoading(false)
       // sync with localstorage
       cacheEnsRecord(address, newRecord)
     }
 
     loadEnsName()
-  }, [address, cache, pending, now, setCache, setPending])
+  }, [address, loading])
 
   if (!address) return
 
-  return cache[address]?.name
+  return ensName
 }
